Coerce unit to string in LRValue.__init__.
The code assumes an instance of LRValue has a .unit, and that
each instance's .unit is of the same type. The problem is that
unit can be a string or a labrad.units.Unit instance. To keep
everything consistent within types.py, we coerce to string in
the LRValue constructor.