Unflatten '*v[]' as l.u.DimensionlessArray.
If a setting advertises 'v' we are allowed to send *v[] or eg. *v[Hz].
This means that that receiving code doesn't know whether to expect
a ValueArray (for *v[Hz]) or an ndarray (for *v[]). Since we have
l.u.DimensionlessArray precisely for this case, we unflatten to that.
Note that l.u.DimensionlessArray subclasses ndarray, so it's unlikely
that this will introduce any bugs or incompatibility.