Fixed packet flattening bug in stream.py
In a recent commit, I changed the LRType subclasses' __flatten__
methods to return a tuple of the flattened string and the LRType
class used for the flattening. The old return signature was just
the flattened string. The problem is that l.stream uses
the (private!) __flatten__ methods on Type classes used to
represent the wire format of the LabRAD packets themselves. So,
in this commit I just added [0] to get the string and drop the
LRType object in those cases.

A much better solution would be to use labrad.types.flatten
everywhere and never use the classes' __flatten__ methods.
For example, instead of

PACKET_TYPE = T.parseTypeTag('(ww)iws')
PACKET_TYPE.__flatten__((context, request, target, data), endianness)[0]

do this

PACKET_TYPE = T.parseTypeTag('(ww)iws')
T.flatten((context, request, target, data), PACKET_TYPE, endianness)

I didn't do this because I don't want to have to write a commit
message with "chomp chomp..." in it. DTS