Add support for TLS connections to labrad
Allows pylabrad clients and servers to connect to the manager
using TLS to secure the connection. The default configuration
uses the 'starttls' mechanism whereby connections start
unencrypted and are subsequently upgraded to be encrypted.
By default, connections to a manager running on the same host
as the client or server will not be encrypted, which is allowed
by the default manager configuration.

As a temporary compatibility measure, clients that fail when
trying to upgrade the connection with starttls will retry with
TLS disabled entirely; this will allow them to connect to older
managers that do not support TLS.

Review: @ejeffrey, @strj