Fix evalLRData to allow None for units and nan/inf for floats.
This is needed for backwards compatibility with data stored in the old
format, since Value and Complex previously allowed None as a unit. In
addition, testing on the existing data stored in the data vault showed
that nan/inf floating point values were not handled correctly, so we
fix those as well.

Fixes #159
Review: @DanielSank