Refactor protocol to share connection logic between server and client
Labrad servers and clients are both network "clients" in the sense that
they make outbound network connections to the manager rather than
accepting inbound connections. By refactoring the labrad server class
to be just a plain python class rather than a twisted ClientFactory,
we can reuse the same connection logic for both labrad clients and
servers. This also lets us write the connection logic in a more direct
style, where previously much of the server connection logic happened
as a side-effect of network callbacks in the Server class which was
an instance of ClientFactory.

Review: @DanielSank