Replace custom Future implementation with standard futures lib.
We use a subclass of Future that is backwards compatible with the old
implementation, but deprecate the .wait() method. In addition, we
change the way Futures are used when calling settings and sending
packets. Previously, these could be given an extra 'wait' argument to
control whether to block until the result was available or to return
a Future. While this usage is still supported, it will now result in
deprecation warnings. Instead, we provide separate .future and
.send_future methods for settings and packets, respectively.

Until all users have transitioned away from the old Future class, we
need to support the addCallback method. This method is awkward because
it requires futures to be mutable (adding a callback changes the value
of that future, rather than producing a new future). We introduce the
MutableFuture class to handle the details of dealing with these
callbacks.

Review: @ejeffrey