Add method to get flattenable representation of a packet.
As we build packets we use type information on settings to flatten data
to types appropriate for those settings. In some situations, we would
like to turn this packet data into a structure that can itself be
flattened and sent via labrad, for example to be forwarded by another
server to the actual target server. In order to do this, we need to be
able to handle flattening structures some of whose parts have already
been flattened.

This change adds a flatten method on LRType which can handle incoming
data that has already been flattened before delegating to the existing
__flatten__ method. This method returns FlatData instead of just a
tuple, which also allows us to simplify the implementation of the
top-level flatten function.

Review: @ejeffrey