Unflatten lists eagerly instead of lazily.
We still use the LazyList class (a subclass of the builtin list)
for unflattened lists that cannot be represented as numpy arrays,
but this is just so that the .aslist, .astuple and .asarray methods
will still work. However, these methods now emit deprecation
warnings, and the LazyList class will be removed in a subsequent
release.

Fixes #227

Review: @ejeffrey