Update node to handle multiple versions of same server.
When scanning for server configs, the node used to take the first
config it encountered for a particular server name and drop any
subsequent configs for the same server name. We change this
behavior so that all configs for a given server name are kept,
but with the version appended to the name for older versions to
keep the names unique. The most recent version keeps its name in
unmodified form.

Review: @ejeffrey