Do not rely on int/long distinction to call manager settings.
The long type no longer exists in python 3, so we want to stop relying
on this type distinction to decide whether to flatten data as i or w.
This comes up when calling manager settings to lookup metadata like
accepted types, since we don't yet have the accepted type info to guide
the flattening.

Review: @ejeffrey