GPIB: Allow refreshing connection to GPIB device manager.
For some GPIB device servers, if the GPIB device manager was started
after the device server, the device server would not update the list of
attached devices once the GPIB device manager was started. Uncomment
out a refresh line in connectToDeviceManager to fix this, so the GPIB
device manager can be started even after GPIB device servers.

Fixes martinisgroup/servers#351