Add ThreadedServer class to handle requests in a thread pool (#294)
Currently, most client-side code uses the synchronous Client class for talking to labrad, which blocks when making requests or returns Futures to allow for firing multiple requests and awaiting the results later. However, server-side code runs in the twisted reactor thread, uses AsyncClient, and must work in non-blocking fashion using twisted Deferreds (or inlineCallbacks). Among other things, this makes it hard to reuse code between servers and clients.

Here we implement a ThreadedServer class that handles requests and other lifecycle callbacks in a thread pool, and makes a synchronous Client available for using the server's labrad connection to make requests. This makes it possible to write server code in the more familiar blocking style and share this code between client and server, while still handling multiple requests concurrently on the server.

Implementing a ThreadedServer requires care to maintain thread safety of any shared data. For situations where concurrent request handling is not needed, we also provide a SingleThreadedServer class that dispatches server lifecycle methods and requests on a single thread. This makes it possible to implement a server without thinking about thread-safety or locks, but only one request can be handled at a time.

Review: @kunalq, @ejeffrey