Clean up ServerProcess implementation (#341)
Changes the way ServerProcess works so that an individual ServerProcess instance cannot be restarted. If we want to restart a server we stop the existing instance and then create and start a new instance. This means that the new instance will pick up newer configuration if the node configs have been refreshed. This also makes it much more straightforward to manage the lifecycle of ServerProcess instances, because once one stops it will never start again, so it can be removed from the node's instances dict and won't be re-added.

This change also lets us clean up the internals of the ServerProcess implementation, in particular the way it communicates with the subprocess, which should make things much more stable. We move the code to listen for labrad server connect instances into ServerProcess itself, so that the node does not have to keep track of instances to notify in the hacky `instances_to_notify` set, and in addition the ServerProcess can listen for messages only during the time when it cares about messages, namely while attempting to startup.

Fixes #303 
Review: @zchen088